# api-gateway/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Asset Marketplace API
  description: |
    REST API documentation for the **Digital Asset Marketplace** — a microservices-based platform
    for buying and selling AI-generated digital assets.

    ## Architecture
    The system consists of three microservices orchestrated through an API Gateway (nginx):
    - **Identity Service** — User registration, authentication, and profile management
    - **Catalog Service** — Digital asset listing, search, and management
    - **Order Service** — Order creation, payment processing, and order tracking

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from the login endpoint.
    Include the token in the `Authorization` header: `Bearer <your_token>`
  version: 1.0.0
  contact:
    name: Digital Asset Marketplace Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local Development (API Gateway)
  - url: http://localhost:3001
    description: Identity Service (Direct)
  - url: http://localhost:3002
    description: Catalog Service (Direct)
  - url: http://localhost:3003
    description: Order Service (Direct)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the Login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "error message"

    # ── Identity Service ──
...
    CreateAssetRequest:
      type: object
      required: [title, price, storage_path]
      properties:
        title:
          type: string
          example: "Neon Cyberpunk City Pack"
        description:
          type: string
          example: "3D assets for cyberpunk themed environments"
        price:
          type: number
          format: double
          minimum: 0.01
          example: 29.99
        category:
          type: string
          example: "3D Models"
        ai_tool_used:
          type: string
          example: "MidJourney"
        resolution:
          type: string
          example: "4096x4096"
        preview_url:
          type: string
          format: uri
          example: "https://example.com/preview.jpg"
        storage_path:
          type: string
          example: "/assets/cyberpunk_pack.zip"
    AssetResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        title:
          type: string
          example: "Neon Cyberpunk City Pack"
        description:
          type: string
          example: "3D assets for cyberpunk themed environments"
        price:
          type: number
          format: double
          example: 29.99
        category:
          type: string
          example: "3D Models"
        ai_tool_used:
          type: string
          example: "MidJourney"
        resolution:
          type: string
          example: "4096x4096"
        preview_url:
          type: string
          example: "https://example.com/preview.jpg"
        sales_count:
          type: integer
          example: 42
        created_at:
          type: string
          format: date-time
          example: "2026-06-06T09:25:05Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-06-06T09:25:05Z"
        created_by:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
...
    UpdateAssetRequest:
      type: object
      required: [title, price, storage_path]
      properties:
        title:
          type: string
          example: "Updated Asset Title"
        description:
          type: string
          example: "Updated description"
        price:
          type: number
          format: double
          minimum: 0.01
          example: 39.99
        ai_tool_used:
          type: string
          example: "DALL-E 3"
        resolution:
          type: string
          example: "8192x8192"
        preview_url:
          type: string
          format: uri
          example: "https://example.com/updated_preview.jpg"
        storage_path:
          type: string
          example: "/assets/updated_pack.zip"

    # ── Order Service ──
...
    SuccessMessage:
      type: object
      properties:
        status:
          type: string
          example: "success"
paths:
...
  /api/assets/:
    post:
      tags: [Assets]
      summary: Create a new digital asset
      description: Creates a new digital asset. Requires authentication (CREATOR role recommended).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '400':
          description: Invalid asset data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assets/{id}:
    get:
      tags: [Assets]
      summary: Get asset by ID
      description: Returns a single asset's details. **This endpoint is public** — no authentication required.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Asset UUID
      responses:
        '200':
          description: Asset details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags: [Assets]
      summary: Update an asset
      description: Updates an existing asset. Only the asset creator can update it.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssetRequest'
      responses:
        '200':
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '403':
          description: Not the asset owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Assets]
      summary: Delete an asset
      description: Deletes an asset. Only the asset creator can delete it.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '403':
          description: Not the asset owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Asset not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/admin/assets:
    get:
      tags: [Admin]
      summary: List all assets
      description: Returns a list of all assets in the marketplace. Requires ADMIN role.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of all assets
        '403':
          description: Forbidden (Not an admin)
  /api/admin/assets/{id}:
    delete:
      tags: [Admin]
      summary: Takedown an asset
      description: Forcibly deletes an asset from the catalog regardless of ownership. Requires ADMIN role.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Asset taken down
        '403':
          description: Forbidden (Not an admin)
        '404':
          description: Asset not found
...